requirements.txt
flask
flask-cors
python-dotenv
requests
pandas
rapidfuzz
serpapi
